Enum cached_packages::aptos_framework_sdk_builder::EntryFunctionCall

pub enum EntryFunctionCall {

    AccountOfferSignerCapability {
        signer_capability_sig_bytes: Vec<u8>,
        account_scheme: u8,
        account_public_key_bytes: Vec<u8>,
        recipient_address: AccountAddress,
    },
    AccountRevokeSignerCapability {
        to_be_revoked_address: AccountAddress,
    },
    AccountRotateAuthenticationKey {
        from_scheme: u8,
        from_public_key_bytes: Vec<u8>,
        to_scheme: u8,
        to_public_key_bytes: Vec<u8>,
        cap_rotate_key: Vec<u8>,
        cap_update_table: Vec<u8>,
    },
    AptosAccountCreateAccount {
        auth_key: AccountAddress,
    },
    AptosAccountTransfer {
        to: AccountAddress,
        amount: u64,
    },
    AptosCoinClaimMintCapability {},
    AptosCoinDelegateMintCapability {
        to: AccountAddress,
    },
    AptosCoinMint {
        dst_addr: AccountAddress,
        amount: u64,
    },
    AptosGovernanceAddApprovedScriptHashScript {
        proposal_id: u64,
    },
    AptosGovernanceCreateProposal {
        stake_pool: AccountAddress,
        execution_hash: Vec<u8>,
        metadata_location: Vec<u8>,
        metadata_hash: Vec<u8>,
    },
    AptosGovernanceVote {
        stake_pool: AccountAddress,
        proposal_id: u64,
        should_pass: bool,
    },
    CodePublishPackageTxn {
        metadata_serialized: Vec<u8>,
        code: Vec<Vec<u8>>,
    },
    CoinTransfer {
        coin_type: TypeTag,
        to: AccountAddress,
        amount: u64,
    },
    CoinUpgradeSupply {
        coin_type: TypeTag,
    },
    ManagedCoinBurn {
        coin_type: TypeTag,
        amount: u64,
    },
    ManagedCoinInitialize {
        coin_type: TypeTag,
        name: Vec<u8>,
        symbol: Vec<u8>,
        decimals: u8,
        monitor_supply: bool,
    },
    ManagedCoinMint {
        coin_type: TypeTag,
        dst_addr: AccountAddress,
        amount: u64,
    },
    ManagedCoinRegister {
        coin_type: TypeTag,
    },
    ResourceAccountCreateResourceAccount {
        seed: Vec<u8>,
        optional_auth_key: Vec<u8>,
    },
    ResourceAccountCreateResourceAccountAndFund {
        seed: Vec<u8>,
        optional_auth_key: Vec<u8>,
        fund_amount: u64,
    },
    ResourceAccountCreateResourceAccountAndPublishPackage {
        seed: Vec<u8>,
        metadata_serialized: Vec<u8>,
        code: Vec<Vec<u8>>,
    },
    StakeAddStake {
        amount: u64,
    },
    StakeIncreaseLockup {},
    StakeInitializeStakeOwner {
        initial_stake_amount: u64,
        operator: AccountAddress,
        voter: AccountAddress,
    },
    StakeInitializeValidator {
        consensus_pubkey: Vec<u8>,
        proof_of_possession: Vec<u8>,
        network_addresses: Vec<u8>,
        fullnode_addresses: Vec<u8>,
    },
    StakeJoinValidatorSet {
        pool_address: AccountAddress,
    },
    StakeLeaveValidatorSet {
        pool_address: AccountAddress,
    },
    StakeReactivateStake {
        amount: u64,
    },
    StakeRotateConsensusKey {
        pool_address: AccountAddress,
        new_consensus_pubkey: Vec<u8>,
        proof_of_possession: Vec<u8>,
    },
    StakeSetDelegatedVoter {
        new_voter: AccountAddress,
    },
    StakeSetOperator {
        new_operator: AccountAddress,
    },
    StakeUnlock {
        amount: u64,
    },
    StakeUpdateNetworkAndFullnodeAddresses {
        pool_address: AccountAddress,
        new_network_addresses: Vec<u8>,
        new_fullnode_addresses: Vec<u8>,
    },
    StakeWithdraw {
        withdraw_amount: u64,
    },
    StakingContractAddStake {
        operator: AccountAddress,
        amount: u64,
    },
    StakingContractCreateStakingContract {
        operator: AccountAddress,
        voter: AccountAddress,
        amount: u64,
        commission_percentage: u64,
        contract_creation_seed: Vec<u8>,
    },
    StakingContractDistribute {
        staker: AccountAddress,
        operator: AccountAddress,
    },
    StakingContractRequestCommission {
        staker: AccountAddress,
        operator: AccountAddress,
    },
    StakingContractResetLockup {
        operator: AccountAddress,
    },
    StakingContractSwitchOperator {
        old_operator: AccountAddress,
        new_operator: AccountAddress,
        new_commission_percentage: u64,
    },
    StakingContractSwitchOperatorWithSameCommission {
        old_operator: AccountAddress,
        new_operator: AccountAddress,
    },
    StakingContractUnlockRewards {
        operator: AccountAddress,
    },
    StakingContractUnlockStake {
        operator: AccountAddress,
        amount: u64,
    },
    StakingContractUpdateVoter {
        operator: AccountAddress,
        new_voter: AccountAddress,
    },
    StakingProxySetOperator {
        old_operator: AccountAddress,
        new_operator: AccountAddress,
    },
    StakingProxySetStakePoolOperator {
        new_operator: AccountAddress,
    },
    StakingProxySetStakePoolVoter {
        new_voter: AccountAddress,
    },
    StakingProxySetStakingContractOperator {
        old_operator: AccountAddress,
        new_operator: AccountAddress,
    },
    StakingProxySetStakingContractVoter {
        operator: AccountAddress,
        new_voter: AccountAddress,
    },
    StakingProxySetVestingContractOperator {
        old_operator: AccountAddress,
        new_operator: AccountAddress,
    },
    StakingProxySetVestingContractVoter {
        operator: AccountAddress,
        new_voter: AccountAddress,
    },
    StakingProxySetVoter {
        operator: AccountAddress,
        new_voter: AccountAddress,
    },
    VersionSetVersion {
        major: u64,
    },
    VestingAdminWithdraw {
        contract_address: AccountAddress,
    },
    VestingDistribute {
        contract_address: AccountAddress,
    },
    VestingResetBeneficiary {
        contract_address: AccountAddress,
        shareholder: AccountAddress,
    },
    VestingResetLockup {
        contract_address: AccountAddress,
    },
    VestingSetBeneficiary {
        contract_address: AccountAddress,
        shareholder: AccountAddress,
        new_beneficiary: AccountAddress,
    },
    VestingSetBeneficiaryResetter {
        contract_address: AccountAddress,
        beneficiary_resetter: AccountAddress,
    },
    VestingSetManagementRole {
        contract_address: AccountAddress,
        role: Vec<u8>,
        role_holder: AccountAddress,
    },
    VestingTerminateVestingContract {
        contract_address: AccountAddress,
    },
    VestingUnlockRewards {
        contract_address: AccountAddress,
    },
    VestingUpdateOperator {
        contract_address: AccountAddress,
        new_operator: AccountAddress,
        commission_percentage: u64,
    },
    VestingUpdateOperatorWithSameCommission {
        contract_address: AccountAddress,
        new_operator: AccountAddress,
    },
    VestingUpdateVoter {
        contract_address: AccountAddress,
        new_voter: AccountAddress,
    },
    VestingVest {
        contract_address: AccountAddress,
    },

}

Structured representation of a call into a known Move entry function.

impl EntryFunctionCall {
    pub fn encode(self) -> TransactionPayload { .. }
    pub fn decode(&TransactionPayload) -> Option<EntryFunctionCall> { .. }
}

Variants

AccountOfferSignerCapability

Fields

signer_capability_sig_bytes: Vec<u8>

account_scheme: u8

account_public_key_bytes: Vec<u8>

recipient_address: AccountAddress

Offers signer capability on behalf of account to the account at address recipient_address. An account can delegate its signer capability to only one other address at one time. signer_capability_key_bytes is the SignerCapabilityOfferProofChallengeV2 signed by the account owner’s key account_scheme is the scheme of the account (ed25519 or multi_ed25519) account_public_key_bytes is the public key of the account owner recipient_address is the address of the recipient of the signer capability - note that if there’s an existing recipient_address in the account owner’s SignerCapabilityOffer, this will replace the previous recipient_address upon successful verification (the previous recipient will no longer have access to the account owner’s signer capability)

AccountRevokeSignerCapability

Fields

to_be_revoked_address: AccountAddress

AccountRotateAuthenticationKey

Fields

from_scheme: u8

from_public_key_bytes: Vec<u8>

to_scheme: u8

to_public_key_bytes: Vec<u8>

cap_rotate_key: Vec<u8>

cap_update_table: Vec<u8>

Generic authentication key rotation function that allows the user to rotate their authentication key from any scheme to any scheme. To authorize the rotation, a signature by the current private key on a valid RotationProofChallenge (cap_rotate_key) demonstrates that the user intends to and has the capability to rotate the authentication key. A signature by the new private key on a valid RotationProofChallenge (cap_update_table) verifies that the user has the capability to update the value at key auth_key on the OriginatingAddress table. from_scheme refers to the scheme of the from_public_key and to_scheme refers to the scheme of the to_public_key. A scheme of 0 refers to an Ed25519 key and a scheme of 1 refers to Multi-Ed25519 keys.

AptosAccountCreateAccount

Fields

auth_key: AccountAddress

Basic account creation methods.

AptosAccountTransfer

Fields

to: AccountAddress

amount: u64

AptosCoinClaimMintCapability

Fields

Only callable in tests and testnets where the core resources account exists. Claim the delegated mint capability and destroy the delegated token.

AptosCoinDelegateMintCapability

Fields

to: AccountAddress

Only callable in tests and testnets where the core resources account exists. Create delegated token for the address so the account could claim MintCapability later.

AptosCoinMint

Fields

dst_addr: AccountAddress

amount: u64

Only callable in tests and testnets where the core resources account exists. Create new coins and deposit them into dst_addr’s account.

AptosGovernanceAddApprovedScriptHashScript

Fields

proposal_id: u64

AptosGovernanceCreateProposal

Fields

stake_pool: AccountAddress

execution_hash: Vec<u8>

metadata_location: Vec<u8>

metadata_hash: Vec<u8>

Create a proposal with the backing stake_pool. @param execution_hash Required. This is the hash of the resolution script. When the proposal is resolved, only the exact script with matching hash can be successfully executed.

AptosGovernanceVote

Fields

stake_pool: AccountAddress

proposal_id: u64

should_pass: bool

Vote on proposal with proposal_id and voting power from stake_pool.

CodePublishPackageTxn

Fields

metadata_serialized: Vec<u8>

code: Vec<Vec<u8>>

Same as publish_package but as an entry function which can be called as a transaction. Because of current restrictions for txn parameters, the metadata needs to be passed in serialized form.

CoinTransfer

Fields

coin_type: TypeTag

to: AccountAddress

amount: u64

Transfers amount of coins CoinType from from to to.

CoinUpgradeSupply

Fields

coin_type: TypeTag

Upgrade total supply to use a parallelizable implementation if it is available.

ManagedCoinBurn

Fields

coin_type: TypeTag

amount: u64

Withdraw an amount of coin CoinType from account and burn it.

ManagedCoinInitialize

Fields

coin_type: TypeTag

name: Vec<u8>

symbol: Vec<u8>

decimals: u8

monitor_supply: bool

Initialize new coin CoinType in Aptos Blockchain. Mint and Burn Capabilities will be stored under account in Capabilities resource.

ManagedCoinMint

Fields

coin_type: TypeTag

dst_addr: AccountAddress

amount: u64

Create new coins CoinType and deposit them into dst_addr’s account.

ManagedCoinRegister

Fields

coin_type: TypeTag

Creating a resource that stores balance of CoinType on user’s account, withdraw and deposit event handlers. Required if user wants to start accepting deposits of CoinType in his account.

ResourceAccountCreateResourceAccount

Fields

seed: Vec<u8>

optional_auth_key: Vec<u8>

Creates a new resource account and rotates the authentication key to either the optional auth key if it is non-empty (though auth keys are 32-bytes) or the source accounts current auth key.

ResourceAccountCreateResourceAccountAndFund

Fields

seed: Vec<u8>

optional_auth_key: Vec<u8>

fund_amount: u64

Creates a new resource account, transfer the amount of coins from the origin to the resource account, and rotates the authentication key to either the optional auth key if it is non-empty (though auth keys are 32-bytes) or the source accounts current auth key. Note, this function adds additional resource ownership to the resource account and should only be used for resource accounts that need access to Coin.

ResourceAccountCreateResourceAccountAndPublishPackage

Fields

seed: Vec<u8>

metadata_serialized: Vec<u8>

code: Vec<Vec<u8>>

Creates a new resource account, publishes the package under this account transaction under this account and leaves the signer cap readily available for pickup.

StakeAddStake

Fields

amount: u64

Add amount of coins from the account owning the StakePool.

StakeIncreaseLockup

Fields

Similar to increase_lockup_with_cap but will use ownership capability from the signing account.

StakeInitializeStakeOwner

Fields

initial_stake_amount: u64

operator: AccountAddress

voter: AccountAddress

Initialize the validator account and give ownership to the signing account except it leaves the ValidatorConfig to be set by another entity. Note: this triggers setting the operator and owner, set it to the account’s address to set later.

StakeInitializeValidator

Fields

consensus_pubkey: Vec<u8>

proof_of_possession: Vec<u8>

network_addresses: Vec<u8>

fullnode_addresses: Vec<u8>

Initialize the validator account and give ownership to the signing account.

StakeJoinValidatorSet

Fields

pool_address: AccountAddress

This can only called by the operator of the validator/staking pool.

StakeLeaveValidatorSet

Fields

pool_address: AccountAddress

Request to have pool_address leave the validator set. The validator is only actually removed from the set when the next epoch starts. The last validator in the set cannot leave. This is an edge case that should never happen as long as the network is still operational.

Can only be called by the operator of the validator/staking pool.

StakeReactivateStake

Fields

amount: u64

Move amount of coins from pending_inactive to active.

StakeRotateConsensusKey

Fields

pool_address: AccountAddress

new_consensus_pubkey: Vec<u8>

proof_of_possession: Vec<u8>

Rotate the consensus key of the validator, it’ll take effect in next epoch.

StakeSetDelegatedVoter

Fields

new_voter: AccountAddress

Allows an owner to change the delegated voter of the stake pool.

StakeSetOperator

Fields

new_operator: AccountAddress

Allows an owner to change the operator of the stake pool.

StakeUnlock

Fields

amount: u64

Similar to unlock_with_cap but will use ownership capability from the signing account.

StakeUpdateNetworkAndFullnodeAddresses

Fields

pool_address: AccountAddress

new_network_addresses: Vec<u8>

new_fullnode_addresses: Vec<u8>

Update the network and full node addresses of the validator. This only takes effect in the next epoch.

StakeWithdraw

Fields

withdraw_amount: u64

Withdraw from account’s inactive stake.

StakingContractAddStake

Fields

operator: AccountAddress

amount: u64

Add more stake to an existing staking contract.

StakingContractCreateStakingContract

Fields

operator: AccountAddress

voter: AccountAddress

amount: u64

commission_percentage: u64

contract_creation_seed: Vec<u8>

Staker can call this function to create a simple staking contract with a specified operator.

StakingContractDistribute

Fields

staker: AccountAddress

operator: AccountAddress

Allow anyone to distribute already unlocked funds. This does not affect reward compounding and therefore does not need to be restricted to just the staker or operator.

StakingContractRequestCommission

Fields

staker: AccountAddress

operator: AccountAddress

Unlock commission amount from the stake pool. Operator needs to wait for the amount to become withdrawable at the end of the stake pool’s lockup period before they can actually can withdraw_commission.

Only staker or operator can call this.

StakingContractResetLockup

Fields

operator: AccountAddress

Convenient function to allow the staker to reset their stake pool’s lockup period to start now.

StakingContractSwitchOperator

Fields

old_operator: AccountAddress

new_operator: AccountAddress

new_commission_percentage: u64

Allows staker to switch operator without going through the lenghthy process to unstake.

StakingContractSwitchOperatorWithSameCommission

Fields

old_operator: AccountAddress

new_operator: AccountAddress

Allows staker to switch operator without going through the lenghthy process to unstake, without resetting commission.

StakingContractUnlockRewards

Fields

operator: AccountAddress

Unlock all accumulated rewards since the last recorded principals.

StakingContractUnlockStake

Fields

operator: AccountAddress

amount: u64

Staker can call this to request withdrawal of part or all of their staking_contract. This also triggers paying commission to the operator for accounting simplicity.

StakingContractUpdateVoter

Fields

operator: AccountAddress

new_voter: AccountAddress

Convenient function to allow the staker to update the voter address in a staking contract they made.

StakingProxySetOperator

Fields

old_operator: AccountAddress

new_operator: AccountAddress

StakingProxySetStakePoolOperator

Fields

new_operator: AccountAddress

StakingProxySetStakePoolVoter

Fields

new_voter: AccountAddress

StakingProxySetStakingContractOperator

Fields

old_operator: AccountAddress

new_operator: AccountAddress

StakingProxySetStakingContractVoter

Fields

operator: AccountAddress

new_voter: AccountAddress

StakingProxySetVestingContractOperator

Fields

old_operator: AccountAddress

new_operator: AccountAddress

StakingProxySetVestingContractVoter

Fields

operator: AccountAddress

new_voter: AccountAddress

StakingProxySetVoter

Fields

operator: AccountAddress

new_voter: AccountAddress

VersionSetVersion

Fields

major: u64

Updates the major version to a larger version. This can be called by on chain governance.

VestingAdminWithdraw

Fields

contract_address: AccountAddress

Withdraw all funds to the preset vesting contract’s withdrawal address. This can only be called if the contract has already been terminated.

VestingDistribute

Fields

contract_address: AccountAddress

Distribute any withdrawable stake from the stake pool.

VestingResetBeneficiary

Fields

contract_address: AccountAddress

shareholder: AccountAddress

Remove the beneficiary for the given shareholder. All distributions will sent directly to the shareholder account.

VestingResetLockup

Fields

contract_address: AccountAddress

VestingSetBeneficiary

Fields

contract_address: AccountAddress

shareholder: AccountAddress

new_beneficiary: AccountAddress

VestingSetBeneficiaryResetter

Fields

contract_address: AccountAddress

beneficiary_resetter: AccountAddress

VestingSetManagementRole

Fields

contract_address: AccountAddress

role: Vec<u8>

role_holder: AccountAddress

VestingTerminateVestingContract

Fields

contract_address: AccountAddress

Terminate the vesting contract and send all funds back to the withdrawal address.

VestingUnlockRewards

Fields

contract_address: AccountAddress

Unlock any accumulated rewards.

VestingUpdateOperator

Fields

contract_address: AccountAddress

new_operator: AccountAddress

commission_percentage: u64

VestingUpdateOperatorWithSameCommission

Fields

contract_address: AccountAddress

new_operator: AccountAddress

VestingUpdateVoter

Fields

contract_address: AccountAddress

new_voter: AccountAddress

VestingVest

Fields

contract_address: AccountAddress

Unlock any vested portion of the grant.

Implementations

impl EntryFunctionCall

pub fn encode(self) -> TransactionPayload

Build an Aptos TransactionPayload from a structured object EntryFunctionCall.

pub fn decode(payload: &TransactionPayload) -> Option<EntryFunctionCall>

Try to recognize an Aptos TransactionPayload and convert it into a structured object EntryFunctionCall.