Function cached_packages::aptos_framework_sdk_builder::account_rotate_authentication_key

pub fn account_rotate_authentication_key(
    from_scheme: u8, 
    from_public_key_bytes: Vec<u8>, 
    to_scheme: u8, 
    to_public_key_bytes: Vec<u8>, 
    cap_rotate_key: Vec<u8>, 
    cap_update_table: Vec<u8>
) -> TransactionPayload

Generic authentication key rotation function that allows the user to rotate their authentication key from any scheme to any scheme. To authorize the rotation, a signature by the current private key on a valid RotationProofChallenge (cap_rotate_key) demonstrates that the user intends to and has the capability to rotate the authentication key. A signature by the new private key on a valid RotationProofChallenge (cap_update_table) verifies that the user has the capability to update the value at key auth_key on the OriginatingAddress table. from_scheme refers to the scheme of the from_public_key and to_scheme refers to the scheme of the to_public_key. A scheme of 0 refers to an Ed25519 key and a scheme of 1 refers to Multi-Ed25519 keys.