Module cached_packages::aptos_token_sdk_builder

Enums

EntryFunctionCall

Structured representation of a call into a known Move entry function.

Functions

token_burn

Burn a token by the token owner

token_burn_by_creator

Burn a token by creator when the token’s BURNABLE_BY_CREATOR is true The token is owned at address owner

token_coin_swap_list_token_for_swap

Token owner lists their token for swapping

token_create_collection_script

create a empty token collection with parameters

token_create_token_script

create token with raw inputs

token_direct_transfer_script

token_initialize_token_script

token_mint_script

Mint more token from an existing token_data. Mint only adds more token to property_version 0

token_mutate_token_properties

mutate the token property and save the new property in TokenStore if the token property_version is 0, we will create a new property_version per token to generate a new token_id per token if the token property_version is not 0, we will just update the propertyMap and use the existing token_id (property_version)

token_opt_in_direct_transfer

token_transfers_cancel_offer_script

token_transfers_claim_script

token_transfers_offer_script