Enum cached_packages::aptos_token_sdk_builder::EntryFunctionCall

pub enum EntryFunctionCall {

    TokenBurn {
        creators_address: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        property_version: u64,
        amount: u64,
    },
    TokenBurnByCreator {
        owner: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        property_version: u64,
        amount: u64,
    },
    TokenCreateCollectionScript {
        name: Vec<u8>,
        description: Vec<u8>,
        uri: Vec<u8>,
        maximum: u64,
        mutate_setting: Vec<bool>,
    },
    TokenCreateTokenScript {

        collection: Vec<u8>,
        name: Vec<u8>,
        description: Vec<u8>,
        balance: u64,
        maximum: u64,
        uri: Vec<u8>,
        royalty_payee_address: AccountAddress,
        royalty_points_denominator: u64,
        royalty_points_numerator: u64,
        mutate_setting: Vec<bool>,
        property_keys: Vec<Vec<u8>>,
        property_values: Vec<Vec<u8>>,
        property_types: Vec<Vec<u8>>,
    
},
    TokenDirectTransferScript {
        creators_address: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        property_version: u64,
        amount: u64,
    },
    TokenInitializeTokenScript {},
    TokenMintScript {
        token_data_address: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        amount: u64,
    },
    TokenMutateTokenProperties {
        token_owner: AccountAddress,
        creator: AccountAddress,
        collection_name: Vec<u8>,
        token_name: Vec<u8>,
        token_property_version: u64,
        amount: u64,
        keys: Vec<Vec<u8>>,
        values: Vec<Vec<u8>>,
        types: Vec<Vec<u8>>,
    },
    TokenOptInDirectTransfer {
        opt_in: bool,
    },
    TokenCoinSwapListTokenForSwap {
        coin_type: TypeTag,
        _creators_address: AccountAddress,
        _collection: Vec<u8>,
        _name: Vec<u8>,
        _property_version: u64,
        _token_amount: u64,
        _min_coin_per_token: u64,
        _locked_until_secs: u64,
    },
    TokenTransfersCancelOfferScript {
        receiver: AccountAddress,
        creator: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        property_version: u64,
    },
    TokenTransfersClaimScript {
        sender: AccountAddress,
        creator: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        property_version: u64,
    },
    TokenTransfersOfferScript {
        receiver: AccountAddress,
        creator: AccountAddress,
        collection: Vec<u8>,
        name: Vec<u8>,
        property_version: u64,
        amount: u64,
    },

}

Structured representation of a call into a known Move entry function.

impl EntryFunctionCall {
    pub fn encode(self) -> TransactionPayload { .. }
    pub fn decode(&TransactionPayload) -> Option<EntryFunctionCall> { .. }
}

Variants

TokenBurn

Fields

creators_address: AccountAddress

collection: Vec<u8>

name: Vec<u8>

property_version: u64

amount: u64

Burn a token by the token owner

TokenBurnByCreator

Fields

owner: AccountAddress

collection: Vec<u8>

name: Vec<u8>

property_version: u64

amount: u64

Burn a token by creator when the token’s BURNABLE_BY_CREATOR is true The token is owned at address owner

TokenCreateCollectionScript

Fields

name: Vec<u8>

description: Vec<u8>

uri: Vec<u8>

maximum: u64

mutate_setting: Vec<bool>

create a empty token collection with parameters

TokenCreateTokenScript

Fields

collection: Vec<u8>

name: Vec<u8>

description: Vec<u8>

balance: u64

maximum: u64

uri: Vec<u8>

royalty_payee_address: AccountAddress

royalty_points_denominator: u64

royalty_points_numerator: u64

mutate_setting: Vec<bool>

property_keys: Vec<Vec<u8>>

property_values: Vec<Vec<u8>>

property_types: Vec<Vec<u8>>

create token with raw inputs

TokenDirectTransferScript

Fields

creators_address: AccountAddress

collection: Vec<u8>

name: Vec<u8>

property_version: u64

amount: u64

TokenInitializeTokenScript

Fields

TokenMintScript

Fields

token_data_address: AccountAddress

collection: Vec<u8>

name: Vec<u8>

amount: u64

Mint more token from an existing token_data. Mint only adds more token to property_version 0

TokenMutateTokenProperties

Fields

token_owner: AccountAddress

creator: AccountAddress

collection_name: Vec<u8>

token_name: Vec<u8>

token_property_version: u64

amount: u64

keys: Vec<Vec<u8>>

values: Vec<Vec<u8>>

types: Vec<Vec<u8>>

mutate the token property and save the new property in TokenStore if the token property_version is 0, we will create a new property_version per token to generate a new token_id per token if the token property_version is not 0, we will just update the propertyMap and use the existing token_id (property_version)

TokenOptInDirectTransfer

Fields

opt_in: bool

TokenCoinSwapListTokenForSwap

Fields

coin_type: TypeTag

_creators_address: AccountAddress

_collection: Vec<u8>

_name: Vec<u8>

_property_version: u64

_token_amount: u64

_min_coin_per_token: u64

_locked_until_secs: u64

Token owner lists their token for swapping

TokenTransfersCancelOfferScript

Fields

receiver: AccountAddress

creator: AccountAddress

collection: Vec<u8>

name: Vec<u8>

property_version: u64

TokenTransfersClaimScript

Fields

sender: AccountAddress

creator: AccountAddress

collection: Vec<u8>

name: Vec<u8>

property_version: u64

TokenTransfersOfferScript

Fields

receiver: AccountAddress

creator: AccountAddress

collection: Vec<u8>

name: Vec<u8>

property_version: u64

amount: u64

Implementations

impl EntryFunctionCall

pub fn encode(self) -> TransactionPayload

Build an Aptos TransactionPayload from a structured object EntryFunctionCall.

pub fn decode(payload: &TransactionPayload) -> Option<EntryFunctionCall>

Try to recognize an Aptos TransactionPayload and convert it into a structured object EntryFunctionCall.